Refactor PKCS7 package and migrate to internal module
We don't want to export the PKCS7 code 3rd parties, therefore moved the PKCS7 implementation to an internal module for better encapsulation and renamed functions to use exportable naming conventions. Updated all references to reflect these changes across the codebase, ensuring consistent usability and improved clarity.
